Cleanup and refactoring preparatory to making root management less of a dumpster fire
